+++ traffic_ops/traffic_ops_golang/traffic_ops_golang.go
@@ -22,6 +22,7 @@ package main
 import (
 	"crypto/tls"
 	"encoding/json"
+	"errors"
 	"flag"
 	"fmt"
 	"io/ioutil"
@@ -64,6 +65,7 @@ func main() {
 	configFileName := flag.String("cfg", "", "The config file path")
 	dbConfigFileName := flag.String("dbcfg", "", "The db config file path")
 	riakConfigFileName := flag.String("riakcfg", "", "The riak config file path (DEPRECATED: use traffic_vault_backend = riak and traffic_vault_config in cdn.conf instead)")
+	backendConfigFileName := flag.String("backendcfg", "", "The backend config file path")
 	flag.Parse()
 
 	if *showVersion {
@@ -164,7 +166,18 @@ func main() {
 		log.Errorln(debugServer.ListenAndServe())
 	}()
 
-	if err := routing.RegisterRoutes(routing.ServerData{DB: db, Config: cfg, Profiling: &profiling, Plugins: plugins, TrafficVault: trafficVault}); err != nil {
+	var backendConfig config.BackendConfig
+	if *backendConfigFileName != "" {
+		backendConfig, err = config.LoadBackendConfig(*backendConfigFileName)
+		routing.SetBackendConfig(backendConfig)
+		if err != nil {
+			log.Errorf("error loading backend config: %v", err)
+		}
+	}
+
+	mux := http.NewServeMux()
+	d := routing.ServerData{DB: db, Config: cfg, Profiling: &profiling, Plugins: plugins, TrafficVault: trafficVault, Mux: mux}
+	if err := routing.RegisterRoutes(d); err != nil {
 		log.Errorf("registering routes: %v\n", err)
 		os.Exit(1)
 	}
@@ -213,7 +226,7 @@ func main() {
 		} else {
 			file.Close()
 		}
-
+		server.Handler = mux
 		if err := server.ListenAndServeTLS(cfg.CertPath, cfg.KeyPath); err != nil {
 			log.Errorf("stopping server: %v\n", err)
 			os.Exit(1)
@@ -232,10 +245,16 @@ func main() {
 		continuousProfile(&profiling, &profilingLocation, cfg.Version)
 	}
 
-	reloadProfilingConfig := func() {
+	reloadProfilingAndBackendConfig := func() {
 		setNewProfilingInfo(*configFileName, &profiling, &profilingLocation, cfg.Version)
+		backendConfig, err = getNewBackendConfig(backendConfigFileName)
+		if err != nil {
+			log.Errorf("could not reload backend config: %v", err)
+		} else {
+			routing.SetBackendConfig(backendConfig)
+		}
 	}
-	signalReloader(unix.SIGHUP, reloadProfilingConfig)
+	signalReloader(unix.SIGHUP, reloadProfilingAndBackendConfig)
 }
 
 func setupTrafficVault(riakConfigFileName string, cfg *config.Config) trafficvault.TrafficVault {
@@ -291,6 +310,19 @@ func setupTrafficVault(riakConfigFileName string, cfg *config.Config) trafficvau
 		return trafficVault
 	}
 	return &disabled.Disabled{}
+}
+
+func getNewBackendConfig(backendConfigFileName *string) (config.BackendConfig, error) {
+	if backendConfigFileName == nil {
+		return config.BackendConfig{}, errors.New("no backend config filename")
+	}
+	log.Infof("setting new backend config to %s", *backendConfigFileName)
+	backendConfig, err := config.LoadBackendConfig(*backendConfigFileName)
+	if err != nil {
+		log.Errorf("error reloading config: %v", err)
+		return backendConfig, err
+	}
+	return backendConfig, nil
 }
 
 func setNewProfilingInfo(configFileName string, currentProfilingEnabled *bool, currentProfilingLocation *string, version string) {